third_party/blink/web_tests/animations/custom-properties/registered-var-dynamic-dependency-expected.txt

This is a testharness.js-based test.
FAIL Animated registered custom properties are invalid at computed-value time when there is a cyclic var() dependency between them. assert_equals: --a at 25% expected "initial-value" but got "150"
Harness: the test ran to completion.
